requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
playwright>=1.30.0
PyPDF2>=3.0.0
pyyaml>=6.0
lxml>=4.9.0
responses>=0.24.0
pytest>=7.0.0
python-telegram-bot>=21.0.0
pytest-mock>=3.0.0
telethon>=1.30.0
